requirements.txt
streamlit
psycopg2-binary
google-genai
langchain
langchain-community
python-dotenv
pypdf
beautifulsoup4
curl-cffi
fastapi
uvicorn[standard]
python-multipart
